Split into datetime and calendar modules
